refactor: swap provider & related hooks to allow for easier extending
