+++ src/contexts/SwapProvider/models.ts
@@ -1,5 +1,11 @@
+import { BehaviorSubject } from 'rxjs';
 import { Address, OptimalRate, PriceString, SwapSide } from '@paraswap/sdk';
-import { WrappedError } from '@src/utils/errors';
+
+import { DestinationInput } from '@src/pages/Swap/utils';
+import { WalletDetails } from '@src/background/services/wallet/models';
+import { Account } from '@src/background/services/accounts/models';
+import { NetworkWithCaipId } from '@src/background/services/network/models';
+import { NetworkFee } from '@src/background/services/networkFee/models';
 
 /**
  * Paraswap may return both data and an error sometimes.
@@ -35,14 +41,39 @@ export const hasParaswapError = (
   return typeof response.error === 'string';
 };
 
-export type SwapParams = {
+export type SwapQuote = OptimalRate | JupiterQuote;
+
+export const isJupiterSwapParams = (
+  swapParams: SwapParams<SwapQuote>,
+): swapParams is SwapParams<JupiterQuote> => {
+  return isJupiterQuote(swapParams.quote);
+};
+
+export const isParaswapSwapParams = (
+  swapParams: SwapParams<SwapQuote>,
+): swapParams is SwapParams<OptimalRate> => {
+  return isParaswapQuote(swapParams.quote);
+};
+
+export const isJupiterQuote = (quote: SwapQuote): quote is JupiterQuote => {
+  return 'inputMint' in quote;
+};
+
+export const isParaswapQuote = (quote: SwapQuote): quote is OptimalRate => {
+  return (
+    'srcAmount' in quote &&
+    'destAmount' in quote &&
+    'srcToken' in quote &&
+    'destToken' in quote
+  );
+};
+
+export type SwapParams<T extends SwapQuote> = {
   srcToken: string;
   destToken: string;
   srcDecimals: number;
   destDecimals: number;
-  srcAmount: string;
-  priceRoute: OptimalRate;
-  destAmount: string;
+  quote: T;
   slippage: number;
 };
 
@@ -53,14 +84,43 @@ export type GetRateParams = {
   destDecimals: number;
   srcAmount: string;
   swapSide?: SwapSide;
+  fromTokenBalance?: bigint;
 };
 
+export type SwapFormValues = {
+  amount?: bigint;
+  toTokenAddress?: string;
+  fromTokenAddress?: string;
+  toTokenDecimals?: number;
+  fromTokenDecimals?: number;
+  destinationInputField?: DestinationInput;
+  fromTokenBalance?: bigint;
+};
+
+type GetRateResult<T extends SwapQuote> =
+  | {
+      error: SwapError;
+      quote: null;
+      destAmount?: string;
+    }
+  | {
+      error: undefined;
+      quote: T;
+      destAmount: string;
+    };
+
 export type SwapContextAPI = {
-  getRate(params: GetRateParams): Promise<{
-    optimalRate: OptimalRate | WrappedError | null;
-    destAmount: string | undefined;
-  }>;
-  swap(params: SwapParams): Promise<void>;
+  getRate(params: GetRateParams): Promise<GetRateResult<SwapQuote>>;
+  swap(params: SwapParams<SwapQuote>): Promise<void>;
+  swapError: SwapError;
+  setSwapError(error: SwapError): void;
+  isSwapLoading: boolean;
+  setIsSwapLoading(isSwapLoading: boolean): void;
+  quote: OptimalRate | JupiterQuote | null;
+  setQuote(quote: OptimalRate | JupiterQuote | null): void;
+  swapFormValuesStream: BehaviorSubject<SwapFormValues>;
+  destAmount: string;
+  setDestAmount(destAmount: string): void;
 };
 
 export const DISALLOWED_SWAP_ASSETS: string[] = [
@@ -111,3 +171,37 @@ export enum SwapErrorCode {
   UnexpectedApiResponse = 'unexpected-api-response',
   CannotBuildTx = 'cannot-build-tx',
 }
+
+export interface SwapError {
+  message?: string;
+  hasTryAgain?: boolean;
+}
+
+export type JupiterQuote = {
+  inputMint: string;
+  inAmount: string;
+  outputMint: string;
+  outAmount: string;
+  otherAmountThreshold: string;
+  swapMode: 'ExactIn' | 'ExactOut';
+  slippageBps: number;
+  platformFee: {
+    amount: string;
+    feeBps: number;
+  };
+  priceImpactPct: string;
+  routePlan: unknown[];
+  contextSlot: number;
+  timeTaken: number;
+};
+
+type WalletState = Partial<{
+  account: Account;
+  network: NetworkWithCaipId;
+  networkFee: NetworkFee | null;
+  walletDetails: WalletDetails;
+}>;
+export type SwapAdapter<T extends SwapQuote> = (walletState: WalletState) => {
+  getRate: (params: GetRateParams) => Promise<GetRateResult<T>>;
+  swap: (params: SwapParams<T>) => Promise<void>;
+};
