+++ src/pages/Swap/models.ts
@@ -1,6 +1,10 @@
 import {
   NetworkTokenWithBalance,
   TokenWithBalanceERC20,
+  TokenWithBalanceSPL,
 } from '@avalabs/vm-module-types';
 
-export type SwappableToken = NetworkTokenWithBalance | TokenWithBalanceERC20;
+export type SwappableToken =
+  | NetworkTokenWithBalance
+  | TokenWithBalanceERC20
+  | TokenWithBalanceSPL;
